# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ImmoPro API",
        "description": "Documentation interactive de l'API ImmoPro — plateforme immobilière.\n\n## Authentification\nLa majorité des endpoints nécessitent un token Bearer (Laravel Sanctum).\nConnectez-vous via **POST /api/login** (Admin/Agent) ou **POST /api/client/login** (Client),\npuis cliquez sur le bouton **Authorize** et collez votre token.\n\n## Rôles\n- **admin** : accès complet à l'administration\n- **agent** : gestion des biens assignés et des visites\n- **client** : propriétaires / acheteurs / locataires",
        "contact": {
            "name": "Support ImmoPro",
            "email": "support@immopro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal de l'API ImmoPro"
        }
    ],
    "paths": {
...
        "/client/login": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Connexion client",
                "description": "Authentifie un utilisateur avec le rôle client. Retourne un token Sanctum.",
                "operationId": "clientLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marie.kone@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MonMotDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|xK3mZ9pQ2rL7nY8sV1jW..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Compte suspendu/bloqué ou email non vérifié"
                    },
                    "422": {
                        "description": "Email ou mot de passe incorrect"
                    }
                }
            }
        },
        "/verify-otp": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Vérification OTP & création du compte (étape 2/2)",
                "description": "Vérifie le code OTP reçu par email et crée le compte client.",
                "operationId": "verifyOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "pending_token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marie.kone@email.com"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "482731",
                                        "maxLength": 6,
                                        "minLength": 6
                                    },
                                    "pending_token": {
                                        "type": "string",
                                        "example": "xK3mZ9pQ2rL7nY8sV1jW4tU6bO0cE5fA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|vR9kP5oN3mQ1eJ8dH..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Code OTP invalide/expiré"
                    }
                }
            }
        },
        "/resend-otp": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Renvoyer un code OTP",
                "description": "Génère et renvoie un nouveau code OTP par email.",
                "operationId": "resendOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marie.kone@email.com"
                                    },
                                    "pending_token": {
                                        "type": "string",
                                        "example": "xK3mZ9pQ2rL7nY8sV1jW4tU6bO0cE5fA",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouveau OTP envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Un nouveau code OTP a été envoyé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune inscription en attente"
                    }
                }
            }
        },
        "/client/me": {
            "get": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Profil du client connecté",
                "operationId": "clientMe",
                "responses": {
                    "200": {
                        "description": "Profil récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/client/logout": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Déconnexion client",
                "operationId": "clientLogout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResource": {
                "title": "Utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@email.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+22501234567",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "Côte d'Ivoire",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "Abidjan",
                        "nullable": true
                    },
                    "profile_picture": {
                        "type": "string",
                        "format": "url",
                        "example": "https://...",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "client",
                        "enum": [
                            "admin",
                            "agent",
                            "client"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "suspended",
                            "blocked"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu après connexion. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}